docs: improve documentation (#2)
This pull request includes updates to the `README.md`, `variables.tf`,
and `outputs.tf` files to enhance the documentation and configuration
descriptions for the Terraform AWS module. The most important changes
include adding a usage example for the knowledge base module, updating
descriptions for various configuration fields, and modifying output
descriptions.

Enhancements to documentation and usage examples:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L9-R40):
Added a detailed usage example for the `knowledge_base` module,
including configuration for `embedding_model`, `storage_type`,
`opensearch_serverless_config`, and `data_source_configurations`.

Updates to configuration descriptions:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L63-R89):
Updated the description for `data_source_configurations` to clarify that
it includes multiple data sources.
*
[`variables.tf`](diffhunk://#diff-05b5a57c136b6ff596500bcbfdcff145ef6cddea2a0e86d184d9daa9a65a288eL106-R106):
Modified the description for `data_source_configurations` to reflect
that it includes configurations for multiple data sources.

Improvements to output descriptions:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L76-R103):
Simplified the descriptions for `data_source_ids` and
`knowledge_base_id` outputs.
*
[`outputs.tf`](diffhunk://#diff-de6c47c2496bd028a84d55ab12d8a4f90174ebfb6544b8b5c7b07a7ee4f27ec7L2-R7):
Updated the descriptions for `knowledge_base_id` and `data_source_ids`
outputs to be more concise.

Consistency in variable descriptions:

*
[`examples/oss-complete/README.md`](diffhunk://#diff-22f732de5b37a5020d3b3adfbb51b58b94d415e887bdd032a27adb12b88d3cf5L80-R80):
Simplified the description for the `creator` input variable.
*
[`examples/oss-complete/variables.tf`](diffhunk://#diff-927797fc01c05018fdccc4f485f700a750b9d5ab2433ebfe4273c681bbb33dbaL4-R4):
Updated the description for the `creator` variable to match the
simplified format.